# src/config/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "ElegancceAPI",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:3333",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Adiciona um produto no bando de dados",
        "description": "",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "modelo de Produto a ser enviado",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Creme para pentear"
                },
                "value": {
                  "type": "number",
                  "example": 45.8
                },
                "description": {
                  "type": "string",
                  "example": "loren ipsum...."
                },
                "qt": {
                  "type": "number",
                  "example": 67
                }
              },
              "required": [
                "nome",
                "value",
                "description",
                "qt"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "product"
        ],
        "summary": "Atualizacao dos dados de um produto",
        "description": "",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "modelo de Produto a ser enviado",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Creme para pentear"
                },
                "value": {
                  "type": "number",
                  "example": 45.8
                },
                "description": {
                  "type": "string",
                  "example": "loren ipsum...."
                },
                "qt": {
                  "type": "number",
                  "example": 67
                }
              },
              "required": [
                "nome",
                "value",
                "description",
                "qt"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/product/{id}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Busca um produto pelo seu id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Remove um produto do banco de dados",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "itenger",
            "description": "Codigo identificador do produto"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/buy": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/discounts": {
      "get": {
        "tags": [
          "discount"
        ],
        "summary": "busca todos os cupons de discontos",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "discount"
        ],
        "summary": "Cria um novo cupon de disconto",
        "description": "",
        "parameters": [
          {
            "name": "",
            "in": "body",
            "description": "modelo de dados do Cupon de disconto",
            "schema": {
              "type": "object",
              "properties": {
                "disocunt": {
                  "type": "number",
                  "example": 50
                },
                "code": {
                  "type": "string",
                  "example": "AUAU"
                },
                "dt_limit": {
                  "type": "number",
                  "example": 1658762344027
                }
              },
              "required": [
                "disocunt",
                "code"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/discounts/{id}": {
      "put": {
        "tags": [
          "discount"
        ],
        "summary": "Atualiza um novo cupon de disconto",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "",
            "in": "body",
            "description": "modelo de dados do Cupon de disconto",
            "schema": {
              "type": "object",
              "properties": {
                "disocunt": {
                  "type": "number",
                  "example": 50
                },
                "code": {
                  "type": "string",
                  "example": "AUAU"
                },
                "dt_limit": {
                  "type": "number",
                  "example": 1658762344027
                }
              },
              "required": [
                "disocunt",
                "code"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "get": {
        "tags": [
          "discount"
        ],
        "summary": "",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "discount"
        ],
        "summary": "Remove o cupon de disconto",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Busca todos os administradores cadastrados no banco de dados",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo identificador de usuario no banco de dados",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Efetua a criação do admin no banco de dados.",
        "description": "",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Modelo de Admin",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Oliver o Adm"
                },
                "email": {
                  "type": "string",
                  "example": "oliver@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "auau123"
                }
              },
              "required": [
                "name",
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Efetua a alteraçao das informaões do admin.",
        "description": "",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Modelo de admin",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 12
                },
                "name": {
                  "type": "string",
                  "example": "Oliver o adm"
                },
                "email": {
                  "type": "string",
                  "example": "oliver@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "auau123"
                }
              },
              "required": [
                "id",
                "name",
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Deleta uma conta de administrador.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}